Fix test compilation errors
* Use legacy AES-KDF mode for KeePass1Reader

+++ tests/TestKeePass1Reader.cpp
@@ -110,7 +110,7 @@ void TestKeePass1Reader::testBasic()
 void TestKeePass1Reader::testMasterKey()
 {
     QVERIFY(m_db->hasKey());
-    QCOMPARE(m_db->kdf()->rounds(), static_cast<quint64>(713));
+    QCOMPARE(m_db->kdf()->rounds(), 713);
 }
 
 void TestKeePass1Reader::testCustomIcons()
